+++ Runtime/Code/Basic/AssetHandling/UiPanelLoadInfo.cs
@@ -1,32 +1,93 @@
-using Codice.Client.BaseCommands.CheckIn;
 using System;
 using UnityEngine;
 
-
 namespace GuiToolkit.AssetHandling
 {
-
+	/// <summary>
+	/// Data container that describes how a UiPanel should be loaded.
+	/// It does not perform any loading itself, but provides all
+	/// necessary information for the AssetLoader / AssetManager.
+	/// </summary>
 	public class UiPanelLoadInfo
 	{
+		/// <summary>
+		/// Defines how the panel instance is managed after creation.
+		/// </summary>
 		public enum EInstantiationType
 		{
-			Instantiate,        // Automatically destroyed after close
-			Pool,               // Pooled
-			InstantiateAndKeep, // Kept in memory after close
+			/// <summary>
+			/// Create new instance, destroy after close.
+			/// </summary>
+			Instantiate,
+
+			/// <summary>
+			/// Reuse pooled instance.
+			/// </summary>
+			Pool,
+
+			/// <summary>
+			/// Create new instance, keep alive after close.
+			/// </summary>
+			InstantiateAndKeep,
 		}
 
-		// Note: The Attributes used here are solely for overview which parts are mandatory
+		// --------------------------------------------------------------------
+		// Note: The attributes [Mandatory] / [Optional] are only markers for
+		//       overview / documentation purposes. They do not enforce rules
+		//       at runtime or compile-time.
+		// --------------------------------------------------------------------
 
+		/// <summary>
+		/// The panel type to be loaded. Must derive from UiPanel.
+		/// </summary>
 		[Mandatory] public Type PanelType;
+
+		/// <summary>
+		/// Optional override for the canonical prefab identifier.
+		/// If null, defaults to the PanelType name.
+		/// </summary>
 		[Optional] public string CanonicalId;
+
+		/// <summary>
+		/// Maximum number of allowed instances. Values &lt;= 0 mean unlimited.
+		/// </summary>
 		[Optional] public int MaxInstances = 0;
+
+		/// <summary>
+		/// Defines how this panel is instantiated (pooled, kept, destroyed).
+		/// Defaults to Pool.
+		/// </summary>
 		[Optional] public EInstantiationType InstantiationType = EInstantiationType.Pool;
+
+		/// <summary>
+		/// Optional parent transform where the panel instance will be attached.
+		/// </summary>
 		[Optional] public Transform Parent = null;
+
+		/// <summary>
+		/// Optional initialization payload passed to the panel on creation.
+		/// </summary>
 		[Optional] public IInitPanelData InitPanelData = null;
+
+		/// <summary>
+		/// Callback invoked on successful load.
+		/// </summary>
 		[Optional] public Action<UiPanel> OnSuccess = null;
+
+		/// <summary>
+		/// Callback invoked when loading fails with an exception.
+		/// </summary>
 		[Optional] public Action<UiPanelLoadInfo, Exception> OnFail = null;
+
+		/// <summary>
+		/// Optional explicit asset provider. If null, the AssetManager will choose one.
+		/// </summary>
 		[Optional] public IAssetProvider AssetProvider = null;
 
+		/// <summary>
+		/// Returns a compact single-line summary of the load info.
+		/// Useful for logs and debugging.
+		/// </summary>
 		public override string ToString()
 		{
 			string parentPath = Parent != null ? Parent.GetPath() : "<null>";
@@ -45,15 +106,16 @@ namespace GuiToolkit.AssetHandling
 				   $"AssetProvider={provider})";
 		}
 
+		/// <summary>
+		/// Returns a multi-line formatted string version of the summary.
+		/// Easier to read in log output when many fields are involved.
+		/// </summary>
 		public string ToMultilineString()
 		{
-			return 
-				ToString()
+			return ToString()
 				.Replace(",", ",\n", StringComparison.Ordinal)
 				.Replace("(", "(\n ", StringComparison.Ordinal)
 				.Replace(")", "\n)\n");
 		}
-
 	}
-
 }
