--- Runtime/Code/Interfaces/AssetHandling.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d0fdf86559964b44481e36e91b6e42a4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
