--- Runtime/Code/Interfaces/AssetHandling/IAssetHandle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6c20775a8b6b6614db4c39695793392c
