--- Runtime/Code/Interfaces/AssetHandling/IInstanceHandle.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-
-namespace GuiToolkit.AssetHandling
-{
-	public interface IInstanceHandle
-	{
-		GameObject Instance { get; }
-		void Release(); // Addressables: ReleaseInstance; Direct: Destroy
-	}
-}
