--- Runtime/Code/Interfaces/AssetHandling/IInstanceHandle.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c5a9d332242a71e4082e09a4d784ffd4
